move current namespace logic to config
